Add intermediate support for GPG 2.0
Turns out 2.0.26 is so buggy it's not possible to make this work.  The
workaround, if one must stick to the 2.0 branch, is to use the brew version of
2.0.30 or greater.

All the changes here are to support that.  There's still work to do to support
both GPG 2.0 and 2.1 since they have difference ways of working.