Document intended usage & implement master key and yubikey reset
The `help` output has been set up as an initial plan for where this is headed.
It is likely to change though.

A `reset-yubikey` target has been created to automatically reset the GPG
applet on a yubikey.

Introduce a hard requirement for GnuPG >= 2.1 as this newer version has fewer
bugs than 2.0 and has added a series of command line flags making it possible
to automate the subkey generation and manipulation.  These tasks were either
not possibe in GnuPG <= 2.0 or required using the cryptic command file
descriptor interface.

Documention for some known caveates that may catch people out.