Fully working yubikey init on GhostBSD
This has been fully tested on a live GhostBSD instance to configure two
Yubikeys.  It worked!  Workarounds have been added for GnuPG's limitation of
cards to which a shadow/stub key can be attributed.  Documentation has been
heavily updated to reflect the changes, and some helper scripts are in place
to facilitate getting secrets and transferable data off the live OS.